[scevent]: allow scevent to clear its cache via messaging system
This is useful to allow external applications writing an external origin
to the database, and associating as additional origin to the seiscomp3
solution of the same event. Without this fix scevent doesn't see new
origin in the database due to the caching of recent events and when trying
to set the new origin as preferred from scolv it doesn't work.