Make playbacks work with past station epochs
Stations are now initialized on the fly based on the pick time of
the first pick received for a given station. Therefore playbacks
will now work with older events and station epochs that ended in the
past. Stations are still set up only once when the first pick is
received, and if the inventory changes during a playback we still
have a problem, but the situation is nevertheless much better than
before. Real-time operation is not affected.