Add files via upload
This is a plugin for the Mindsets SeismicPi HAT board (https://mindsetsonline.co.uk/shop/seismicpi/). It is based on the minilogger plugin but features selection between single or multi channel configuration and internal (accelerometer) or external inputs. When using external inputs it can be used with up to 4 channels. The sample rate is fixed at 32sps. 
The station configuration options for the SeismicPI board are defined in ~/seiscomp/var/lib/seedlink/<NETWORK>.<STATION>.ini file. 
In the setup_configruation directory are included two optional files to automate station configuration: SETTINGS.TXT, where SeismicPI configuration options are defined and UPDATE.pl which updates the seiscomp with the new configurations and start it.